requirements.txt
dlib==19.17.0
numpy==1.16.2
Pillow==5.4.1
nudged==0.3.1
scipy>=0.17.0
matplotlib>=2.0.0,!=3.0.0
networkx>=2.0
imageio>=2.0.1
PyWavelets>=0.4.0
Cython >= 0.23.4
